backend/microservices/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
httpx>=0.27.0,<1.0.0
Pillow>=11.0.0,<12.0.0
